Allow users not to flatten module installation directory structure

+++ Prism/Functions/Install-PrivateModule.ps1
@@ -115,7 +115,7 @@ function Install-PrivateModule
 
             $moduleDirPath = Join-Path -Path $installDirPath -ChildPath $module.Name
             Write-Debug "moduleDirPath        ${moduleDirPath}"
-            if ($singleVersion -and (Test-Path -Path $moduleDirPath))
+            if ($singleVersion -and $Configuration.FlattenModules -and (Test-Path -Path $moduleDirPath))
             {
                 Write-Debug "Removing ${moduleDirPath}"
                 Remove-Item -Path $moduleDirPath -Recurse -Force
@@ -139,7 +139,7 @@ function Install-PrivateModule
 
             # Windows has a 260 character limit for path length. Reduce paths by removing extraneous version
             # directories.
-            if ($singleVersion)
+            if ($singleVersion -and $Configuration.FlattenModules)
             {
                 $modulePath = Join-Path -Path $installDirPath -ChildPath $module.name
                 $versionDirName = $module.version
